refactor: move `RouteUtils.kt` to `admin-common`
EPMDJ-10877
