+++ admin-auth/src/test/kotlin/com/epam/drill/admin/auth/TestUtils.kt
@@ -23,7 +23,7 @@ import com.epam.drill.admin.auth.config.CLAIM_USER_ID
 import com.epam.drill.admin.auth.config.AuthDatabaseConfig
 import com.epam.drill.admin.auth.entity.ApiKeyEntity
 import com.epam.drill.admin.auth.entity.UserEntity
-import com.epam.drill.admin.auth.model.DataResponse
+import com.epam.drill.admin.common.model.DataResponse
 import com.epam.drill.admin.auth.principal.Role
 import com.epam.drill.admin.auth.principal.User
 import com.zaxxer.hikari.HikariConfig
